Backend/requirements.txt
alembic==1.8.1
apturl==0.5.2
asgiref==3.5.2
async-generator==1.10
asyncio==3.4.3
attrs==22.2.0
Authlib==0.15.5
azure-common==1.1.28
azure-core==1.26.0
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
bcrypt==3.2.0
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brlapi==0.8.3
Brotli==1.0.9
certifi==2022.12.7
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
crispy-bootstrap5==0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.0.5
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==22.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.2.1
duplicity==0.8.21
email-validator==1.3.0
eventlet==0.33.0
exceptiongroup==1.1.0
fasteners==0.14.1
filelock==3.8.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
fyers-api==1.0.9
fyers-apiv2==2.0.5
greenlet==1.1.3.post0
gyp==0.1
h11==0.14.0
httpagentparser==1.9.5
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
lockfile==0.12.2
logger==1.4
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msal==1.20.0
msal-extensions==0.3.1
msrest==0.7.1
netifaces==0.11.0
node==1.0
numpy==1.24.1
oauthlib==3.2.0
odict==1.9.0
olefile==0.46
outcome==1.2.0
packaging==22.0
pandas==1.5.2
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
pgadmin4==6.15
Pillow==9.0.1
platformdirs==2.5.4
plumber==1.7
portalocker==2.6.0
protobuf==3.12.4
psutil==5.9.3
psycopg-binary==3.1.8
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyotp==2.7.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
SecretStorage==3.3.1
selenium==4.7.2
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speaklater3==1.4
SQLAlchemy==1.4.42
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
tornado==6.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
ua-parser==0.16.1
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
user-agents==2.2.0
virtualenv==20.16.7
wadllib==1.3.6
webdriver-manager==3.8.5
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.1.2
wsproto==1.2.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.4
zope.proxy==4.5.0
